backend/requirements.txt
fastapi
uvicorn
pymongo
mysql-connector-python
langchain
langchain_community
openai
pydantic
pandas
json5
